1. All stuff to be generated moved to templates directory.\n 2. Using nunjucks as templating engine.\n 3. Generated sample project: .\n 4. Fixed bug related to  being used  module: 'Error [ERR_REQUIRE_ESM]: require() of ES Module /Users/nitinbansal/dev/github/notsc/node_modules/inquirer/lib/index.js from /Users/nitinbansal/dev/github/notsc/dist/generator.js not supported.'
