+++ package.json
@@ -40,8 +40,9 @@
   },
   "devDependencies": {
     "@changesets/cli": "^2.29.5",
-    "@types/inquirer": "^9.0.7",
+    "@types/inquirer": "^8.2.10",
     "@types/node": "^20.10.0",
+    "@types/nunjucks": "^3.2.6",
     "nodemon": "^3.1.10",
     "prettier": "^3.6.2",
     "ts-node": "^10.9.2",
@@ -51,7 +52,10 @@
     "@types/fs-extra": "^11.0.4",
     "chalk": "^4.1.2",
     "commander": "^11.1.0",
+    "express": "^5.1.0",
     "fs-extra": "^11.3.0",
-    "inquirer": "^9.2.12"
+    "inquirer": "^8.2.6",
+    "nunjucks": "^3.2.4",
+    "supertest": "^7.1.3"
   }
 }
