+++ templates/.env.example
@@ -1,0 +1,10 @@
+PORT=3000
+NODE_ENV=development
+{% if database %}
+MONGODB_URI=mongodb://localhost:27017/{{ projectName }}
+{% endif %}{% if redis %}
+REDIS_HOST=localhost
+REDIS_PORT=6379
+REDIS_PASSWORD=
+REDIS_DB=0
+{% endif %}
