+++ templates/.gitignore
@@ -1,0 +1,78 @@
+node_modules/
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+
+dist/
+build/
+*.tsbuildinfo
+
+.env
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+logs/
+*.log
+
+pids
+*.pid
+*.seed
+*.pid.lock
+
+coverage/
+*.lcov
+
+.nyc_output
+
+jspm_packages/
+
+.npm
+
+.eslintcache
+
+.node_repl_history
+
+*.tgz
+
+.yarn-integrity
+
+.env.test
+
+.cache
+.parcel-cache
+
+.next
+
+.nuxt
+
+.vuepress/dist
+
+.serverless/
+
+.fusebox/
+
+.dynamodb/
+
+.tern-port
+
+.vscode-test
+
+.DS_Store
+.DS_Store?
+._*
+.Spotlight-V100
+.Trashes
+ehthumbs.db
+Thumbs.db
+
+.vscode/
+.idea/
+*.swp
+*.swo
+*~
+
+tmp/
+temp/
