+++ templates/docker-compose.yml
@@ -1,0 +1,31 @@
+version: '3.8'
+
+services:
+  app:
+    build: .
+    ports:
+      - "3000:3000"
+    environment:
+      - NODE_ENV=development
+      - PORT=3000
+    {% if database or redis %}depends_on:
+      {% if database %}- mongodb
+      {% endif %}{% if redis %}- redis
+      {% endif %}{% endif %}
+{% if database %}
+  mongodb:
+    image: mongo:latest
+    ports:
+      - "27017:27017"
+    volumes:
+      - mongodb_data:/data/db
+{% endif %}{% if redis %}
+  redis:
+    image: redis:alpine
+    ports:
+      - "6379:6379"
+{% endif %}
+{% if database %}
+volumes:
+  mongodb_data:
+{% endif %}
