+++ templates/nodemon.json
@@ -1,0 +1,6 @@
+{
+  "watch": ["src"],
+  "ext": "ts,js",
+  "ignore": ["src/**/*.test.ts"],
+  "exec": "ts-node src/server.ts"
+}
