+++ templates/src/middlewares/errorHandler.ts
@@ -1,0 +1,27 @@
+import { Request, Response, NextFunction } from 'express';
+import { errorResponse } from '../utils/api-response';
+
+export interface AppError extends Error {
+  statusCode?: number;
+  isOperational?: boolean;
+}
+
+export const errorHandler = (
+  err: AppError,
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
+  const statusCode = err.statusCode || 500;
+  const message = err.message || 'Internal Server Error';
+
+  console.error('Error:', {
+    message: err.message,
+    stack: err.stack,
+    url: req.url,
+    method: req.method,
+    ip: req.ip
+  });
+
+  errorResponse(res, statusCode, message);
+};
