+++ templates/src/routes/healthRoutes.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import { getHealth } from '../controllers/healthController';
+
+const router = Router();
+
+router.get('/health', getHealth);
+
+export default router;
