+++ templates/src/routes/index.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import healthRoutes from './healthRoutes';
+
+const router = Router();
+
+router.use('/', healthRoutes);
+
+export default router;
