+++ templates/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './health';
+export * from '../utils/api-response';
